% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npMarginal.R
\name{npMarginal}
\alias{npMarginal}
\title{Marginal feature screening under the Neyman-Pearson paradigm}
\usage{
npMarginal(x, y, method = c("logistic", "penlog", "svm", "randomforest",
  "lda", "slda", "nb", "nnb", "ada", "tree"),
  p.adjust.methods = c("holm", "hochberg", "hommel", "bonferroni", "BH",
  "BY", "fdr", "none"), N, alpha, delta = 0.05, epsilon = 0.05,
  l0 = 0.5, l1 = 0.5, seed = NULL, ncores = detectCores() - 1, ...)
}
\arguments{
\item{x}{a design matrix}

\item{y}{a vector containing binary labels 0 and 1}

\item{method}{base classification method
\itemize{
  \item logistic: Logistic regression. \code{\link[stats]{glm}} function with family = 'binomial'
  \item penlog: Penalized logistic regression with LASSO penalty. \code{\link[glmnet]{glmnet}} in glmnet package
  \item svm: Support Vector Machines. \code{\link[e1071]{svm}} in e1071 package
  \item randomforest: Random Forest. \code{\link[randomForest]{randomForest}} in randomForest package
  \item lda: Linear Discriminant Analysis. \code{\link[MASS]{lda}} in MASS package
  \item slda: Sparse Linear Discriminant Analysis with LASSO penalty.
  \item nb: Naive Bayes. \code{\link[e1071]{naiveBayes}} in e1071 package
  \item nnb: Nonparametric Naive Bayes. \code{\link[naivebayes]{naive_bayes}} in naivebayes package
  \item ada: Ada-Boost. \code{\link[ada]{ada}} in ada package
}}

\item{p.adjust.methods}{multiple testing adjustment method. See \code{\link[stats]{p.adjust.methods}}}

\item{N}{a positive integer indicating the maximum number of marginal features to be kept}

\item{alpha}{a numeric scalar between 0 and 1 indicating the population type I error control}

\item{delta}{a numeric scalar between 0 and 1 indicating the violation rate. Default: 0.05}

\item{epsilon}{a numeric scalar between 0 and 1 indicating the significance level. Default: 0.05}

\item{l0}{a numeric scalar between 0 and 1 indicating the proportion of leave-out class 0 data points. Default: 0.5}

\item{l1}{a numeric scalar between 0 and 1 indicating the proportion of leave-out class 1 data points. Default: 0.5}

\item{seed}{random seed}

\item{ncores}{a positive integer that specifies the number of cores for computing. Default: number of cores - 1.}

\item{...}{additional argument for base classification methods.}
}
\value{
\code{npMarginal} returns a list with the following components:
\item{alpha}{user-specified population type I error control}
\item{delta}{user-specified violation rate}
\item{epsilon}{user-specified significance level}
\item{N}{user-specified maximum feature number}
\item{pval.unadj}{a vector of unadjusted p-values}
\item{pval.adj}{a vector of adjusted p-values}
\item{p.adjust.methods}{user-specified p-value adjusting method}
\item{feature}{features that pass marginal feature screening}
}
\description{
Marginal feature screening under the Neyman-Pearson paradigm
}
\examples{
### example 1 ###
n = 1000; p = 1000
y = rbinom(n,size = 1, p =0.5)
x = matrix(NA, nrow =n,ncol =p)
x1 = rmvnorm(sum(y==1), mean = rep(1, p), sigma = diag(1,p))
x0 = rmvnorm(sum(y==0), mean = seq(from = -5, to = 4, length.out = p), sigma = diag(1,p))
x[y==1,] = x1
x[y==0,] = x0
table(y)
temp1 = npMarginal(x,y,method = "logistic",
N = 50,
p.adjust.methods = 'BH',
alpha =  0.05,
epsilon = 0.1,
l0 = 0.5,
l1 = 0.5)
plot(temp1$pval.unadj, ylim = c(0,1))
abline(h = epsilon)

### example 2 ###
n = 1000; p = 1000
y = rbinom(n,size = 1, p =0.5)
x = matrix(NA, nrow =n,ncol =p)
x1 = rmvnorm(sum(y==1), mean = rep(1, p), sigma = diag(1,p))
x0 = rmvnorm(sum(y==0), mean = seq(from = -2, to = 1, length.out = p), sigma = diag(1,p))
x[y==1,] = x1
x[y==0,] = x0
table(y)
temp2 = npMarginal(x,y,method = "logistic",
N = 50,
p.adjust.methods = 'BH',
alpha =  0.05,
epsilon = 0.1,
l0 = 0.5,
l1 = 0.5)
plot(temp2$pval.unadj, ylim = c(0,1))
abline(h = epsilon)

}
\references{
\bold{FILL HERE}
}
\author{
Yiling Chen, \email{yiling0210@ucla.edu}
}
